Выполнено задание 9 спринта

+++ src/tracker/tasks/Subtask.java
@@ -6,7 +6,7 @@ public class Subtask extends AbstractTask implements IntersectableTask {
     private Epic topEpic;
 
     public Subtask(String name, String description, TaskStatus taskStatus,
-                   ZonedDateTime startTime, int durationOfMinutes) {
+                   ZonedDateTime startTime, long durationOfMinutes) {
         super(name, description, taskStatus, startTime, durationOfMinutes);
         this.taskType = TaskType.SUBTASK;
     }
