+++ src/tracker/tasks/Task.java
@@ -4,7 +4,7 @@ import java.time.ZonedDateTime;
 
 public class Task extends AbstractTask implements IntersectableTask {
     public Task(String name, String description, TaskStatus taskStatus,
-                ZonedDateTime startTime, int durationOfMinutes) {
+                ZonedDateTime startTime, long durationOfMinutes) {
         super(name, description, taskStatus, startTime, durationOfMinutes);
         this.taskType = TaskType.TASK;
     }
